#914 adopt wedo interpreter files for new communication protocol
