+++ OpenRobertaParent/OpenRobertaServer/staticResources/js/app/wedoInterpreter/native.js
@@ -22,8 +22,8 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
             case "gyro":
                 sensorName = "tiltsensor";
                 break;
-            case "button":
-                sensorName = "buttons";
+            case "buttons":
+                sensorName = "button";
                 break;
             case C.TIMER:
                 return timerGet(port); // RETURN timer value
@@ -55,8 +55,7 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
         name = WEDO.getBrickIdByName(name);
         var robotText = 'robot: ' + name + ', port: ' + port;
         U.p(robotText + ' led on color ' + color);
-        var op = { 'type': 'command', 'actuator': 'light', 'device': name, 'color': color };
-        var cmd = { 'target': 'wedo', 'op': op };
+        var cmd = { 'target': 'wedo', 'type': 'command', 'actuator': 'light', 'device': name, 'color': color };
         WEBVIEW_C.jsToAppInterface(cmd);
     }
     exports.ledOnAction = ledOnAction;
@@ -64,8 +63,7 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
         name = WEDO.getBrickIdByName(name);
         var robotText = 'robot: ' + name + ', port: ' + port;
         U.p(robotText + ' led off');
-        var op = { 'type': 'command', 'actuator': 'light', 'device': name, 'color': 0 };
-        var cmd = { 'target': 'wedo', 'op': op };
+        var cmd = { 'target': 'wedo', 'type': 'command', 'actuator': 'light', 'device': name, 'color': 0 };
         WEBVIEW_C.jsToAppInterface(cmd);
     }
     exports.statusLightOffAction = statusLightOffAction;
@@ -73,8 +71,7 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
         name = WEDO.getBrickIdByName(name); // TODO: better style
         var robotText = 'robot: ' + name;
         U.p(robotText + ' piezo: ' + ', frequency: ' + frequency + ', duration: ' + duration);
-        var op = { 'type': 'command', 'actuator': 'piezo', 'device': name, 'frequency': frequency, 'duration': duration };
-        var cmd = { 'target': 'wedo', 'op': op };
+        var cmd = { 'target': 'wedo', 'type': 'command', 'actuator': 'piezo', 'device': name, 'frequency': frequency, 'duration': duration };
         WEBVIEW_C.jsToAppInterface(cmd);
     }
     exports.toneAction = toneAction;
@@ -83,8 +80,7 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
         var robotText = 'robot: ' + name + ', port: ' + port;
         var durText = duration === -1 ? ' w.o. duration' : (' for ' + duration + ' msec');
         U.p(robotText + ' motor speed ' + speed + durText);
-        var op = { 'type': 'command', 'actuator': 'motor', 'device': name, 'action': 'on', 'id': port, 'direction': speed < 0 ? 1 : 0, 'power': Math.abs(speed) };
-        var cmd = { 'target': 'wedo', 'op': op };
+        var cmd = { 'target': 'wedo', 'type': 'command', 'actuator': 'motor', 'device': name, 'action': 'on', 'id': port, 'direction': speed < 0 ? 1 : 0, 'power': Math.abs(speed) };
         WEBVIEW_C.jsToAppInterface(cmd);
     }
     exports.motorOnAction = motorOnAction;
@@ -92,8 +88,7 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
         name = WEDO.getBrickIdByName(name); // TODO: better style
         var robotText = 'robot: ' + name + ', port: ' + port;
         U.p(robotText + ' motor stop');
-        var op = { 'type': 'command', 'actuator': 'motor', 'device': name, 'action': 'stop', 'id': port };
-        var cmd = { 'target': 'wedo', 'op': op };
+        var cmd = { 'target': 'wedo', 'type': 'command', 'actuator': 'motor', 'device': name, 'action': 'stop', 'id': port };
         WEBVIEW_C.jsToAppInterface(cmd);
     }
     exports.motorStopAction = motorStopAction;
@@ -116,4 +111,3 @@ define([ 'exports', 'constants.interpreter', 'state.interpreter', 'util.interpre
     }
     exports.close = close;
 });
-
